requirements.txt
opencv-python
numpy
imutils
pytesseract
Pillow
streamlit